requirements.txt
discord.py-self
colorama
asyncio
requests
httpx
py
pyfiglet
googletrans==4.0.0-rc1